requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
tflite-runtime==2.14.0
numpy==1.24.2
pillow==11.0.0